More refactorings
